# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Xet Server API
  version: 0.8.0
  description: |
    HTTP APIs exposed by `xetd`: a wire-compatible Xet CAS server
    (`internal/casserver`, mounted at `/v1` and `/v2`), an optional Hub API
    shim (`internal/hubserver`, mounted on a separate port via `-hub-addr`)
    that lets the real `hf` CLI talk to this server end-to-end, and the Xet
    Data API (`internal/api`, mounted at `/v1` alongside the CAS protocol,
    on a disjoint set of literal paths) - a simple, non-wire-compatible
    chunk/dedup API for quick manual testing via the `xet` CLI.

    ## Authentication

    Every endpoint except telemetry/storage-stats (CAS) and `/v1/stats`
    (Xet Data API) is gated by a pluggable `auth.Authenticator`,
    defaulting to no enforcement at all (this server's pre-v0.8.0
    behavior). Enabled via `xetd -auth-token <secret>` (or
    `$XETD_AUTH_TOKEN`/`$HF_TOKEN`), which installs a single shared bearer
    token checked against every request:

    - **401 Unauthorized** - no `Authorization` header, or a token that
      fails validation entirely (wrong token, malformed header).
    - **403 Forbidden** - a valid token that lacks the scope (`read` or
      `write`) the endpoint requires.

    In this UI, use **Authorize** and paste the same secret the server was
    started with; it is sent as `Authorization: Bearer <token>` on every
    "Try it out" call. If the server has no auth configured you can skip
...

    ## Trying endpoints from this page

...

...

...

...
servers:
...
security:
  - bearerAuth: []
components:
  responses:
    Unauthorized:
      description: No credential, or one that fails validation entirely.
      content:
        text/plain:
          schema: { type: string }
    Forbidden:
      description: Valid credential, but lacking the scope this endpoint requires.
      content:
        text/plain:
          schema: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Optional. Only enforced if xetd was started with -auth-token (or
        $XETD_AUTH_TOKEN/$HF_TOKEN) set to something other than "None". A
        server with no auth configured accepts every request regardless of
        whether this header is present.
  schemas:
    ErrorResponse:
      type: object
      description: >-
        casserver's plain-text error body (Content-Type: text/plain via
        http.Error) is NOT this shape - only hubserver and the Xet Data
        API's JSON error paths use this. See each endpoint's response for
        its actual error content-type.
      properties:
        error:
          type: string
      required: [error]
...
    CreateRepoRequest:
      type: object
      properties:
        name: { type: string }
        organization: { type: string }
        type: { type: string, description: 'Defaults to "model" if empty.' }
      required: [name]
    CreateRepoResponse:
      type: object
      properties:
        url: { type: string, description: 'Resolved repo ID, e.g. "org/name".' }
      required: [url]
...
    LFSBatchRequest:
      type: object
      description: >-
        git-LFS batch request body. Only operation and objects are read;
        clients also send transfers/hash_algo/ref, which are ignored.
      properties:
        operation:
          type: string
          description: Only "upload" is accepted; anything else is a 400.
          example: upload
        transfers:
          type: array
          description: >-
            Transfer types the client supports. Informational here - this
            server always answers "xet" - but a client that did not offer
            "xet" ignores the choice and falls back to plain LFS.
          items: { type: string }
          example: [basic, multipart, xet]
        objects:
          type: array
          description: >-
            One entry per file being negotiated. Capped per batch; a larger
            array is rejected with 413 rather than truncated, since
            silently dropping entries would leave the client believing
            files were negotiated that never were. Real clients batch 256
            (huggingface_hub's UPLOAD_BATCH_MAX_NUM_FILES).
          items: { $ref: '#/components/schemas/LFSBatchObject' }
      required: [operation, objects]
    LFSBatchObject:
      type: object
      description: One file's git-LFS identity - its plain SHA-256 and byte size.
      properties:
        oid: { type: string, description: 'Plain SHA-256 of the file (hex), NOT the Xet/Merkle hash.' }
        size: { type: integer, format: int64 }
      required: [oid, size]
    LFSBatchResponse:
      type: object
      description: >-
        Batch response selecting the Xet transfer. Objects are echoed
        back unchanged; per-object `actions` are omitted deliberately (see
        the endpoint description).
      properties:
        transfer:
          type: string
          enum: [xet]
          description: Always "xet" - this server stores files only as Xet xorbs.
        objects:
          type: array
          items: { $ref: '#/components/schemas/LFSBatchObject' }
      required: [transfer, objects]
...
paths:
...
  /api/repos/create:
    post:
      tags: [Hub]
      summary: Create (or no-op re-touch) a repo
      description: Requires write scope. Idempotent - repeat calls with the same name are no-ops.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRepoRequest' }
      responses:
        '200':
          description: Repo exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateRepoResponse' }
        '400':
          description: Invalid body or missing name.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /{repo_id}.git/info/lfs/objects/batch:
    post:
      tags: [Hub]
      summary: git-LFS batch negotiation (entry point to the Xet upload path)
      description: >-
        Requires write scope. Despite the git-LFS URL, this is how a real
        client enters the XET upload path: huggingface_hub offers the
        server ["basic", "multipart", "xet"] and branches on the transfer
        it picks, so answering "xet" here is what routes the client into
        _upload_xet_files (xet-write-token + hf_xet upload to CAS). A
        server without this endpoint gets no Xet upload traffic at all -
        the client falls back to plain LFS and fails. Files below
        huggingface_hub's ~5 MB LFS threshold are committed inline and
        never reach this endpoint.


        Per-object `actions` are deliberately absent from the response:
        _validate_batch_actions requires only oid+size, and the Xet path
        ignores actions entirely, re-deriving everything from the
        xet-write-token plus the local file.


        Accepts the same optional repo-type prefix as resolve
        (datasets/..., spaces/...). The body must be exactly one JSON
        value - trailing data is rejected rather than silently ignored -
        and is bounded in both byte size and object count, since the
        response echoes every object back.
      parameters:
        - { name: repo_id, in: path, required: true, schema: { type: string }, example: alice/my-model }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LFSBatchRequest' }
      responses:
        '200':
          description: Transfer negotiated as "xet"; objects echoed back.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LFSBatchResponse' }
        '400':
          description: >-
            Malformed JSON, trailing data after the JSON value, a
            non-"upload" operation, or a repo_id whose namespace/name is
            empty.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '405':
          description: Method other than POST.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
        '413':
          description: Body over the size cap, or more objects than the per-batch cap.
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...